refactor: update response to only respond feat: add namespace process to process the request
