+++ src/response/response.cpp
@@ -1,10 +1,12 @@
 #include "File.hpp"
 #include "Page.hpp"
+#include "Text.hpp"
 #include "code.hpp"
 #include "Connection.hpp"
 #include "header.hpp"
 #include "logger.hpp"
 #include "parser.hpp"
+#include "process.hpp"
 #include "response.hpp"
 #include "status.hpp"
 #include <iostream>
@@ -13,210 +15,30 @@
 
 using namespace std;
 
-bool response::isDirectory(const std::string &path) {
-
-	struct stat info;
-
-	if (stat(path.c_str(), &info) == 0)
-		return (info.st_mode & S_IFDIR) != 0;
-
-	return false;
-}
-
-bool response::isFile(const std::string &path) {
-
-	struct stat info;
-
-	if (stat(path.c_str(), &info) == 0)
-		return (info.st_mode & S_IFREG) != 0;
-
-	return false;
-}
-
-// TODO: refactor
-bool response::isCGI(const std::string &path) {
-	size_t pos = path.find_last_of(".");
-	if (pos == std::string::npos) {
-		return false;
-	}
-	std::string extension = path.substr(pos);
-	return extension == ".php" || extension == ".py" || extension == ".go";
-}
-
-string response::getFileExtension(string path) {
-
-	list<string> splited_path = parser::split(path, '/');
-	if (splited_path.empty())
-		return "";
-
-	size_t pos;
-	list<string>::iterator it = splited_path.begin();
-	for (; it != splited_path.end(); it++) {
-
-		if (*it == "." || *it == "..")
-			continue;
-
-		pos = it->find_first_of(".");
-		if (pos == string::npos)
-			continue;
-
-		return it->substr(pos + 1);
-	}
-
-	return "";
-}
-
-//Location response::isPathValid(Connection *connection) {
-//	string path = connection->getPath();
-//	if (path.find("..") != string::npos || path.find('/') == string::npos) {
-//		return Location();
-//	}
-//
-//	while (!path.empty()) {
-//		Location temp = connection->getServer().getLocationByURI(path);
-//		if (not temp.getURI().empty())
-//		{
-//			return temp;
-//		}
-//
-//		size_t lastSlash = path.find_last_of('/');
-//		if (lastSlash == 0)
-//			path = "/";
-//		else if (lastSlash != string::npos)
-//			path = path.substr(0, lastSlash);
-//		else
-//			path.clear();
-//	}
-//	return Location();
-//}
-
-Location isValidPath(Connection *connection) {
-
-	list<Location> locations;
-	locations.push_back(connection->getServer().getLocationByURI("/"));
-
-	string path = connection->getPath();
-
-	size_t pos = path.find_first_of("?");
-	if (pos != string::npos)
-		path.erase(pos);
-
-	list<string> splited_path = parser::split(path, '/');
-	if (splited_path.empty())
-		return locations.back();
-
-	string tmp;
-
-	list<string>::iterator it = splited_path.begin();
-	for (; it != splited_path.end(); it ++) {
-
-		cout << *it << endl;
-
-		if (*it == ".")
-			continue;
-
-		if (*it == "..") {
-			if (locations.size() > 1)
-				locations.pop_back();
-
-			continue;
-		}
-
-		if (it->find_first_of(".?") != string::npos) {
-
-			return locations.back();
-		}
-
-		tmp += "/" + *it;
-
-		Location location = connection->getServer().getLocationByURI(tmp);
-		if (!location.empty())
-			locations.push_back(location);
-	}
-
-	cout << "location: " << tmp << endl;
-
-	return locations.back();
-}
-
 static void buildHeaderAndBody(Connection *connection) {
 
 	string header_connection = (*connection)[header::CONNECTION];
+	string header_location = (*connection)[header::LOCATION];
 
 	connection->setProtocol(response::PROTOCOL);
 	connection->setHeaders(response::EMPTY_HEADER);
 
 	connection->addHeader(header::CONNECTION, header_connection);
+	connection->addHeader(header::LOCATION, header_location);
 
 	connection->buildResponse();
 	connection->setSend(true);
 }
 
-bool response::checkIndex(const Location &location, Connection *connection) {
-	const std::set<std::string> &indexes = location.getIndexes();
-	const string &path = connection->getPath();
-	const string &bar = path.find_last_of('/') == path.size() - 1 ? "" : "/";
-
-	typedef std::set<std::string>::const_iterator set_iterator;
-	for (set_iterator it = indexes.begin(); it != indexes.end(); ++it) {
-        std::string indexPath = path + bar + *it;
-        if (isFile(indexPath)) {
-            connection->setPath(indexPath);
-            return true;
-        }
-    }
-    return false;
-}
-
-string getPathFromReferer(string url) {
-
-	size_t pos = url.find_first_of("://");
-	if (pos != string::npos)
-		url.erase(0, pos + 3);
-
-	pos = url.find_first_of("/");
-	if (pos != string::npos)
-		url.erase(0, pos);
-
-	pos = url.find_first_of("?");
-	if (pos != string::npos)
-		url.erase(pos);
-
-	return url;
-}
-
 void response::pageOK(Connection *connection) {
-
-	string url = (*connection)[header::REFERER];
-	if (url.size())
-		connection->setPath(getPathFromReferer(url) + connection->getPath());
-
-	Location location = isValidPath(connection);
-	cout << location << endl;
-	if (location.empty())
-		return response::pageNotFound(connection);
-
-	string queryString = connection->getPath().find('?') != string::npos ? connection->getPath().substr(connection->getPath().find('?')) : "";
-	if (not queryString.empty())
-		connection->setQueryString(queryString);
-
-	string path = location.getRoot() + connection->getPath();
-	path = not queryString.empty() ? path.substr(0, path.find('?')) : path;
-	connection->setPath(path);
-	if (isDirectory(path) && not checkIndex(location, connection))
-		return response::pageForbbiden(connection);
-
-	//cout << path << " " << getFileExtension(connection->getPath()) << endl;
-
-	if (isCGI(path))
-	{
-		// cout << "CGI Path:::::: " << path << endl;
-		// function to handle CGI
-	}
 
 	connection->setCode(code::OK);
 	connection->setStatus(status::OK);
-	connection->setFile(new File(connection->getPath()));
+
+	process::request(connection);
+	if (connection->getCode() != code::OK)
+		return;
+
 	buildHeaderAndBody(connection);
 
 	logger::info(connection->getHost() + " "
@@ -225,6 +47,16 @@ void response::pageOK(Connection *connection) {
 		+ connection->getProtocol() + " "
 		+ connection->getCode() + " - "
 		+ connection->getHeaderByKey(header::USER_AGENT));
+}
+
+void response::pageMovedPermanently(Connection *connection) {
+
+	connection->setCode(code::MOVED_PERMANENTLY);
+	connection->setStatus(status::MOVED_PERMANENTLY);
+	// TODO: check if getReturnURI is a valid path or a text
+	//connection->addHeader(header::LOCATION, location.getReturnURI());
+	//connection->setFile(new Text(connection->getLocation().getReturnURI()));
+	buildHeaderAndBody(connection);
 }
 
 void response::pageBadRequest(Connection *connection) {
